vurderingForretningssted.test.jsx enzyme rewrite
